catalogues_app/requirements.txt
Django==2.1.3
djangorestframework==3.9.0
SQLAlchemy==1.2.13
SQLAlchemy-Utils==0.33.6
psycopg2==2.7.5
